Rename hooksx to harnessopts
